Unimodal baseline code
Notebook for fine-tuning a set of NLP baselines (unimodal) using only Subtask 1 data. Untested models are too large to upload but are available from the OneDrive link in the SemEval document (PDF on main page) under 'unimodel baselines'. Code is modified version of the SemEval 3 task on the same techniques - will be removed if the repo is public and only for testing purposes/starting for NLP models in PyTorch.